+++ tests/svlog/operators.sv
@@ -1,0 +1,16 @@
+module A;
+	int a, b, c;
+	initial begin
+		c = a && b;
+		c = a || b;
+		c = a & b;
+		c = a ~& b;
+		c = a | b;
+		c = a ~| b;
+		c = a ^ b;
+		c = a ~^ b;
+		c = a ^~ b;
+	end
+endmodule
+
+//@ elab A