+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Support for typedefs.
 - Support for port assignments in instantiations.
 - Support for continuous assignments in modules.
+- Support for bitwise logic operators.
 
 ### Fixed
 - Fix assigning values to output ports.